requirements.txt
flask
celery
redis
feedparser
newspaper3k
moviepy
requests
sentence-transformers
peewee
pyyaml
schedule
colorlog
python-dotenv
